pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "tabular_augmentation"
version = "0.0.17"
authors = [
    { name = "zhuxiaofei", email = "dylan@zhxfei.com" },
]
description = "Implementing easy-to-use methods for classical and novel tabular data augmentation and synthesis."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "ctgan>=0.7.4",
    "torchvision",
    "sdv>=1.3.0",
    "tensorboard>=2.13.0",
    "imbalanced-learn>=0.11.0",
    "scikit-learn>=1.0.2",
    "ema-pytorch",
    "xgboost==1.7.6",
    "category-encoders",
    "icecream",
    "catboost",
    "libzero==0.0.8",
    "tomli==1.2.2",
    "tomli-w==0.4.0",
    "optuna==2.10.1",
    "tqdm"
]

[project.urls]
"Homepage" = "https://github.com/zhxfei/tabular_augmentation"
"Bug Tracker" = "https://github.com/zhxfei/tabular_augmentation/issues"
